@@ -2,7 +2,10 @@ import re, httpx, asyncio
 
 from nonebot import on_keyword
 from nonebot.rule import Rule
-from nonebot.adapters.onebot.v11 import Event, Message
+from nonebot.adapters.onebot.v11 import (
+    MessageEvent,
+    Message
+)
 from nonebot.exception import ActionFailed
 
 from .filter import is_not_in_disable_group
@@ -11,15 +14,13 @@ from ..data_source.ytdlp import get_video_info, ytdlp_download_video
 from ..config import *
 
 
-tiktok = on_keyword(keywords={"tiktok.com", "vt.tiktok.com", "vm.tiktok.com"}, rule = Rule(is_not_in_disable_group))
+tiktok = on_keyword(
+    keywords={"tiktok.com", "vt.tiktok.com", "vm.tiktok.com"},
+    rule = Rule(is_not_in_disable_group)
+)
 
 @tiktok.handle()
-async def _(event: Event) -> None:
-    """
-        tiktok解析
-    :param event:
-    :return:
-    """
+async def _(event: MessageEvent):
     # 消息
     message: str = event.message.extract_plain_text().strip()
     url_reg = r"(http:|https:)\/\/(www|vt|vm).tiktok.com\/[A-Za-z\d._?%&+\-=\/#@]*"
@@ -35,8 +36,12 @@ async def _(event: Event) -> None:
         url = str(resp.url)
     elif prefix == "vm":
         async with httpx.AsyncClient() as client:
-            resp = await client.get(url, headers={ "User-Agent": "facebookexternalhit/1.1" }, follow_redirects=True,
-                              proxies=PROXY)
+            resp = await client.get(
+                url,
+                headers={ "User-Agent": "facebookexternalhit/1.1" },
+                follow_redirects=True,
+                proxies=PROXY
+            )
         url = str(resp.url)
     try:
         info = await get_video_info(url)
@@ -44,8 +49,8 @@ async def _(event: Event) -> None:
     except Exception as e:
         await tiktok.send(f"{NICKNAME}解析 | TikTok - 标题获取出错: {e}")
     try:
-        video_name = await ytdlp_download_video(url = url)
-        await tiktok.send(await get_video_seg(video_name))
+        video_path = await ytdlp_download_video(url = url)
+        await tiktok.send(await get_video_seg(video_path))
     except Exception as e:
         if not isinstance(e, ActionFailed):
             await tiktok.send(f"下载失败 | {e}")
